% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findcutCox.R
\name{findcutCox}
\alias{findcutCox}
\title{findcutCox}
\usage{
findcutCox(
  target,
  event,
  time,
  confound = NULL,
  numcut,
  initial_rr = NULL,
  initial_cut = NULL,
  initial_domain = NULL,
  numgen,
  gap = NULL
)
}
\description{
find the optimal location of cutpoints for a continuous variable
}
\examples{
## example for no initial values
findcutCox(BMI,Death,Death_surtime,stage3,numcut=3,initial_rr=NULL,initial_cut=NULL,initial_domain=NULL,numgen=15,gap=NULL)
## example for giving initial values
findcutCox(BMI, Death, Death_surtime, stage3, numcut=3, initial_rr=initial_rr,initial_cut=initial_cut, initial_domain=initial_domain, numgen=15, gap=NULL)
## example for no initial values (numcut=2，repercenting finding the best 2 cutoffs)
findcutCox(BMI,Death,Death_surtime,confound=NULL,numcut=2,initial_rr=NULL,initial_cut=NULL,initial_domain=NULL,numgen=15,gap=NULL)
## example for no initial values (numcut=2，repercenting finding the best 2 cutoffs)
findcutCox(BMI,Death, Death_surtime, confound=NULL, numcut=2, initial_rr=initial_rr,initial_cut=initial_cut,initial_domain=initial_domain,numgen=15,gap=NULL)
}
