% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Findcutnum.R
\name{findcutnum}
\alias{findcutnum}
\title{findcutnum}
\usage{
findcutnum(factor, outcome, datatype, nmin = 20, segment = 100)
}
\description{
find the optimal number of cut-off points for a continuous risk factor
}
\examples{
## Analyzing data from survival information
findcutnum(factor=BMI, outcome=cbind(event,OS), datatype="survival", nmin=5, segment=100)
## Analyzing data from survival information
findcutnum(factor=invasion, outcome=LVSI, datatype="logistic", nmin=5, segment=100)
}
